Refactor and handling database migrations
