+++ alembic/versions/85028f013e6d_zero_migration_revision.py
@@ -1,0 +1,30 @@
+"""Zero Migration revision
+
+Revision ID: 85028f013e6d
+Revises: 28099038d8d6
+Create Date: 2024-09-30 14:48:49.349114
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+import bookworm
+
+# revision identifiers, used by Alembic.
+revision: str = '85028f013e6d'
+down_revision: Union[str, None] = '28099038d8d6'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
